feat(dev): expose Metro bundler over Tailscale
Resolves the laptop's Tailscale MagicDNS name and feeds it to Expo via
REACT_NATIVE_PACKAGER_HOSTNAME so physical devices on the Tailnet can
reach the Metro bundler. Always binds --host lan unless the user passed
their own --host. Falls back silently when Tailscale isn't installed or
logged in. Extracts the existing web hostname resolver into a shared
helper so the web dev server and mobile dev script use the same logic.
