+++ packages/web/scripts/dev-with-tailscale.ts
@@ -1,96 +1,15 @@
-import { execFileSync, spawn } from 'node:child_process';
+import { spawn } from 'node:child_process';
 import { existsSync, readFileSync } from 'node:fs';
 import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
+import { resolveTailscaleHostname } from '../../../scripts/lib/tailscale-hostname';
+
 const scriptDirectory = dirname(fileURLToPath(import.meta.url));
 const rootDirectory = resolve(scriptDirectory, '../../..');
 
-type HostResolution = {
-  hostname: string;
-  source: 'env' | 'tailscale' | 'fallback';
-  reason?: string;
-};
-
 const DEFAULT_WEB_PORT = '3000';
 const DEFAULT_BACKEND_PORT = '8080';
-const TAILSCALE_STATUS_TIMEOUT_MS = 1500;
-
-function normalizeHostname(value: string): string | null {
-  const trimmed = value.trim().replace(/\.$/, '');
-  if (!trimmed) return null;
-
-  if (trimmed.includes('://') || trimmed.includes('/') || trimmed.includes(':')) {
-    return null;
-  }
-
-  if (!/^[a-zA-Z0-9.-]+$/.test(trimmed)) {
-    return null;
-  }
-
-  return trimmed.toLowerCase();
-}
-
-function resolveHostname(): HostResolution {
-  const envHostnameRaw = process.env.TAILSCALE_HOSTNAME;
-  if (envHostnameRaw !== undefined) {
-    const envHostname = normalizeHostname(envHostnameRaw);
-    if (envHostname) {
-      return { hostname: envHostname, source: 'env' };
-    }
-
-    return {
-      hostname: 'localhost',
-      source: 'fallback',
-      reason: 'TAILSCALE_HOSTNAME is invalid; falling back to localhost',
-    };
-  }
-
-  try {
-    const statusJson = execFileSync('tailscale', ['status', '--json'], {
-      encoding: 'utf8',
-      timeout: TAILSCALE_STATUS_TIMEOUT_MS,
-      stdio: ['ignore', 'pipe', 'pipe'],
-    });
-
-    const parsed = JSON.parse(statusJson) as { Self?: { DNSName?: string } };
-    const dnsName = parsed.Self?.DNSName;
-
-    if (!dnsName) {
-      return {
-        hostname: 'localhost',
-        source: 'fallback',
-        reason: 'tailscale status missing Self.DNSName; falling back to localhost',
-      };
-    }
-
-    const normalizedDnsName = normalizeHostname(dnsName);
-    if (!normalizedDnsName) {
-      return {
-        hostname: 'localhost',
-        source: 'fallback',
-        reason: 'tailscale DNSName invalid; falling back to localhost',
-      };
-    }
-
-    return { hostname: normalizedDnsName, source: 'tailscale' };
-  } catch (error) {
-    const errorCode = (error as NodeJS.ErrnoException).code;
-    if (errorCode === 'ENOENT') {
-      return {
-        hostname: 'localhost',
-        source: 'fallback',
-        reason: 'tailscale CLI not installed; falling back to localhost',
-      };
-    }
-
-    return {
-      hostname: 'localhost',
-      source: 'fallback',
-      reason: 'tailscale unavailable; falling back to localhost',
-    };
-  }
-}
 
 function setDefaultEnv(key: string, value: string): void {
   if (!process.env[key]) {
@@ -133,7 +52,7 @@ function main(): void {
 
   const webPort = process.env.PORT || DEFAULT_WEB_PORT;
   const backendPort = process.env.BACKEND_PORT || DEFAULT_BACKEND_PORT;
-  const resolution = resolveHostname();
+  const resolution = resolveTailscaleHostname();
 
   // HTTPS mode: orchestrator has provisioned a Tailscale cert and injected
   // DEV_HTTPS_CERT_FILE / DEV_HTTPS_KEY_FILE. Both must be present to switch
